+++ src/color.rs
@@ -30,7 +30,7 @@ pub const fn rgba(r: u8, g: u8, b: u8, a: u8) -> RgbaColor {
     RgbaColor { r, g, b, a }
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy)]
+#[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq)]
 pub struct RgbColor {
     pub r: u8,
     pub g: u8,