+++ src/gui/top_panel.rs
@@ -2,6 +2,7 @@ use {
     super::{dialogs::LuaColorDialog, message_dialog::Icon, top_menu::top_menu, Gui},
     crate::{
         app::App,
+        color::RgbColor,
         util::human_size,
         value_color::{ColorMethod, Palette},
     },
@@ -60,8 +61,8 @@ pub fn ui(ui: &mut Ui, gui: &mut Gui, app: &mut App, lua: &Lua, font_size: u16, 
                         );
                         ui.selectable_value(
                             &mut presentation.color_method,
-                            ColorMethod::Mono,
-                            ColorMethod::Mono.name(),
+                            ColorMethod::Mono(RgbColor::WHITE),
+                            ColorMethod::Mono(RgbColor::WHITE).name(),
                         );
                         ui.selectable_value(
                             &mut presentation.color_method,
@@ -100,6 +101,12 @@ pub fn ui(ui: &mut Ui, gui: &mut Gui, app: &mut App, lua: &Lua, font_size: u16, 
                     });
                 ui.color_edit_button_rgb(&mut app.preferences.bg_color);
                 ui.label("Bg color");
+                if let ColorMethod::Mono(color) = &mut presentation.color_method {
+                    let mut rgb = [color.r, color.g, color.b];
+                    ui.color_edit_button_srgb(&mut rgb);
+                    [color.r, color.g, color.b] = rgb;
+                    ui.label("Text color");
+                }
                 if let ColorMethod::Custom(arr) = &mut presentation.color_method {
                     let Some(&byte) = app.data.get(app.edit_state.cursor) else {
                         return;